agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a]>=1.27.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
